Better cache management in SPI and USART drivers.
- More fine-grained cache management in the SPI and USART drivers.
- Fixed a potential deadlock in the USART driver due to throwing an exception with IRQ disabled.
- Better diagnosis when a thread terminate from an uncaught C++ exception.
- Removed some unneeded calls to __DSB().

+++ Include/Kernel/Drivers/STM32/SPIDriver.h
@@ -57,7 +57,7 @@ struct SPIDriverSetup
     int32_t                     CRCPolynomial       = 7;
     int32_t                     WordSize            = 8;
     uint32_t                    InterWordIdleCycles = 0;
-    uint32_t                    ReceiveBufferSize   = 0;
+    size_t                      ReceiveBufferSize   = 0;
 };
 
 class SPIDriverINode : public KINode
